2019/requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
docstring-to-markdown==0.10
idna==3.4
jedi==0.18.1
parso==0.8.3
pluggy==1.0.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.6.0
PyYAML==6.0
requests==2.28.1
ujson==5.5.0
urllib3==1.26.12
